+++ src/main/java/shadows/apotheosis/adventure/boss/BossItemManager.java
@@ -1,10 +1,5 @@
 package shadows.apotheosis.adventure.boss;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import java.util.Random;
-
 import javax.annotation.Nullable;
 
 import com.google.gson.Gson;
@@ -12,7 +7,6 @@ import com.google.gson.GsonBuilder;
 
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.util.random.WeightedRandom;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.phys.AABB;
 import net.minecraftforge.coremod.api.ASMAPI;
@@ -23,11 +17,11 @@ import shadows.apotheosis.util.EntityTypeDeserializer;
 import shadows.apotheosis.util.GearSet.SetPredicate;
 import shadows.apotheosis.util.GearSet.SetPredicateAdapter;
 import shadows.placebo.json.NBTAdapter;
-import shadows.placebo.json.PlaceboJsonReloadListener;
 import shadows.placebo.json.RandomAttributeModifier;
 import shadows.placebo.json.SerializerBuilder;
+import shadows.placebo.json.WeightedJsonReloadListener;
 
-public class BossItemManager extends PlaceboJsonReloadListener<BossItem> {
+public class BossItemManager extends WeightedJsonReloadListener<BossItem> {
 
 	//Formatter::off
 	public static final Gson GSON = new GsonBuilder()
@@ -45,38 +39,8 @@ public class BossItemManager extends PlaceboJsonReloadListener<BossItem> {
 
 	public static final BossItemManager INSTANCE = new BossItemManager();
 
-	private final List<BossItem> entries = new ArrayList<>();
-	private volatile int weight = 0;
-
 	public BossItemManager() {
 		super(AdventureModule.LOGGER, "bosses", false, false);
-	}
-
-	@Override
-	protected void beginReload() {
-		super.beginReload();
-		this.entries.clear();
-	}
-
-	@Override
-	protected void register(ResourceLocation id, BossItem item) {
-		if (!this.registry.containsKey(id)) {
-			this.registry.put(id, item);
-			this.entries.add(item);
-		} else AdventureModule.LOGGER.error("Attempted to register a boss item with name {}, but it already exists!", id);
-	}
-
-	@Override
-	protected void onReload() {
-		super.onReload();
-		if (this.entries.size() == 0) throw new RuntimeException("No Bosses were registered.  This is not supported.");
-		Collections.shuffle(this.entries);
-		this.weight = WeightedRandom.getTotalWeight(this.entries);
-		if (this.weight == 0) throw new RuntimeException("The total boss weight is zero.  This is not supported.");
-	}
-
-	public BossItem getRandomItem(Random rand) {
-		return WeightedRandom.getRandomItem(rand, this.entries, this.weight).get();
 	}
 
 	@Nullable